move TextInput model to models/legacy
